+++ src/Core/FunctionLayer/Camera.h
@@ -17,7 +17,7 @@ class Camera
 {
 public:
 	// @brief generate a ray by NDC coord ranges between [-1,1].
-	Ray generateRay(Point2f NDC);
+	Ray generateRay(Point2d NDC);
 
 	// TODO sampler?
 };
